Extract date parsing logic into utils class
